machine-learning/services/video_processing/requirements.txt
pydantic
aiokafka
moviepy
numpy
pillow
tensorflow
uvicorn
